разделён код на отдельные js-модули
